Autogen for region read, cloud scheduler examples
